chore: unified config
