--- frontend/gameConfig/gamePlayConfig.js
@@ -1,18 +1,0 @@
-export const LEADERBOARD_CONFIG = {
-  MAX_LENGTH_TOUCH: 3,
-  MAX_LENGTH_DESKTOP: 10,
-};
-
-export const DEATH_POPUP_CONFIG = {
-  INACTIVITY_DELAY_MS: 30000,
-  COUNTDOWN_SECONDS: 10,
-  REDIRECT_URL: "web.html",
-};
-
-export const SPEEDUP_CONFIG = {
-  DURATION_MS: 5000,
-};
-
-export const UI_MESSAGES = {
-  RESET_SKIN_HINT: "Click 'R' to reset skin!",
-};
