--- frontend/gameConfig/inputConfig.js
@@ -1,21 +1,0 @@
-export const INPUT_CODES = {
-  FEED: "KeyW",
-  SPLIT: "Space",
-  MASK_SKIN: "KeyA",
-  RESET_SKIN: "KeyR",
-  CLEAR_KEYWORD: "Backspace",
-  TOGGLE_PAUSE: "Escape",
-};
-
-export const INPUT_KEYS = {
-  SPEEDUP: "Shift",
-};
-
-export const DEV_KEYWORDS = {
-  EXPORT_JSON: "logs1",
-  EXPORT_TEXT: "logs2",
-};
-
-export const MESSAGES = {
-  SKIN_COPY_TOOLTIP: "Click 'A' to apply the new skin!",
-};
