--- frontend/gameConfig/localStorageKeys.js
@@ -1,8 +1,0 @@
-export const LOCAL_STORAGE_KEYS = {
-  PLAYER_NAME: "playerName",
-  AVAILABLE_SKINS: "availableSkins",
-  LAST_SCORE: "lastScore",
-  MODE: "gameMode",
-  PRIVATE_ROOM: "privateRoomId",
-  CUSTOM_SKIN: "customSkin",
-};
