--- frontend/gameConfig/networkConfig.js
@@ -1,21 +1,0 @@
-export const LOCAL_HOSTNAMES = ["localhost", "127.0.0.1"];
-export const LOCAL_PREFIXES = ["192.168."];
-
-export const LOCAL_PORT = 8080;
-
-export const PRODUCTION_WS_URL = "ws://161.35.75.14:8080/ws";
-
-export const RECONNECT_DELAY = 3000;
-
-export const DEFAULTS = {
-  GAME_MODE: "FFA",
-};
-
-export const MESSAGE_TYPES = {
-  JOIN: "join",
-  INPUT: "input",
-  SPLIT: "split",
-  FEED: "feed",
-  SPEEDUP: "speedup",
-  LEAVE: "leave",
-};
