--- frontend/gameConfig/rendererConfig.js
@@ -1,75 +1,0 @@
-export const RENDER_CONFIG = {
-  GRID: {
-    SIZE: 50,
-    LINE_COLOR: "rgb(15, 66, 85)",
-    LINE_WIDTH: 1,
-  },
-
-  TEXT: {
-    MIN_FONT_SIZE: 10,
-    MAX_FONT_SIZE: 24,
-    FONT_SIZE_RATIO: 0.8,
-    FONT_FAMILY: "Inter",
-    PADDING: 4,
-    BACKGROUND_COLOR: "rgba(0, 0, 0, 0.5)",
-    DEFAULT_COLOR: "white",
-  },
-
-  BLOB: {
-    DEFAULT_WOBBLE_INTENSITY: 0.04,
-    DEFAULT_WOBBLE_SPEED: 0.008,
-    DEFAULT_MIN_POINTS: 8,
-    DEFAULT_MAX_POINTS: 16,
-    DEFAULT_POINT_DENSITY: 10,
-    BORDER_ALPHA: 0.3,
-    BORDER_WIDTH_RATIO: 0.02,
-    MIN_BORDER_WIDTH: 1,
-
-    BREATHE_SPEED: 0.003,
-    BREATHE_AMPLITUDE: 0.02,
-    MULTIPLY_ALPHA: 0.1,
-    SCREEN_ALPHA: 0.05,
-  },
-
-  PLAYER_BLOB: {
-    WOBBLE_INTENSITY: 0.04,
-    WOBBLE_SPEED: 0.008,
-    MIN_POINTS: 8,
-    MAX_POINTS: 16,
-    POINT_DENSITY: 10,
-    BREATHE_EFFECT: true,
-  },
-
-  BUSH_BLOB: {
-    HIDDEN_VISIBILITY: 30,
-    VISIBLE_VISIBILITY: 100,
-    BORDER_COLOR: "#336633",
-    WOBBLE_INTENSITY: 0.02,
-    WOBBLE_SPEED: 0.004,
-    MIN_POINTS: 12,
-    MAX_POINTS: 20,
-    POINT_DENSITY: 8,
-  },
-
-  TRAIL: {
-    LENGTH: 5,
-    MAX_ALPHA: 0.3,
-    MIN_SIZE_RATIO: 0.5,
-    OFFSET_MULTIPLIER: 1.5,
-  },
-
-  OUTLINE: {
-    DANGER_COLOR: "#ff4444",
-    SUCCESS_COLOR: "#44ff44",
-    ALPHA: 0.4,
-    DURATION: 2000,
-  },
-
-  BACKGROUND_COLOR: "#111111",
-
-  FOOD: {
-    DEFAULT_VISIBILITY: 100,
-  },
-
-  CANVAS_ID: "gameCanvas",
-};
