--- frontend/gameConfig/uiConfig.js
@@ -1,26 +1,0 @@
-export const SPEED_BAR_CONFIG = {
-  MAX_SEGMENTS: 5,
-};
-
-export const ERROR_POPUP_CONFIG = {
-  TIMEOUT_MS: 2000,
-  CLASS_VISIBLE: "visible",
-};
-
-export const UI_MESSAGES = {
-  ROOM_ID_COPIED: "Room ID copied to clipboard!",
-};
-
-export const ELEMENT_IDS = {
-  PLAYER_NAME: "playerName",
-  PLAYER_SCORE: "playerScore",
-  LEADERBOARD: "leaderboardList",
-  CANCEL_EXIT: "cancelExit",
-  CONFIRM_EXIT: "confirmExit",
-  EXIT_POPUP: "exitPopup",
-  DEATH_POPUP: "deathPopup",
-  FINAL_SCORE: "finalScore",
-  ERROR_POPUP: "errorPopup",
-  SPEED_BAR: "speedBarFill",
-  ROOM_ID: "roomId",
-};
