--- frontend/gameConfig/zoomConfig.js
@@ -1,11 +1,0 @@
-export const ZOOM_CONFIG = {
-  MIN_RADIUS: 20,
-  MAX_RADIUS: 500,
-  MIN_COVERAGE: 1 / 9,
-  MAX_COVERAGE: 3 / 4,
-  ZOOM_STEP: 0.1,
-  MANUAL_ZOOM_THRESHOLD: 100,
-  MIN_MANUAL_SCALE: 1,
-  MAX_MANUAL_SCALE: 4,
-  SMOOTHNESS: 0.1,
-};
