+++ frontend/gamePage.html
@@ -25,6 +25,7 @@
           <div id="roomId" class="copyable" title="Click to copy room ID">
             <span>Copy Room ID:</span
             ><svg
+              class="svg"
               width="16"
               height="16"
               viewBox="0 0 23 23"
@@ -32,9 +33,10 @@
               xmlns="http://www.w3.org/2000/svg"
             >
               <path
-                d="M17 17H19.5C20.6046 17 21.5 16.1046 21.5 15V3C21.5 1.89543 20.6046 1 19.5 1H7.5C6.39543 1 5.5 1.89543 5.5 3V6M3 22H15C16.1046 22 17 21.1046 17 20V8C17 6.89543 16.1046 6 15 6H3C1.89543 6 1 8V20C1 21.1046 1.89543 22 3 22Z"
+                d="M17 17H19.5C20.6046 17 21.5 16.1046 21.5 15V3C21.5 1.89543 20.6046 1 19.5 1H7.5C6.39543 1 5.5 1.89543 5.5 3V6M3 22H15C16.1046 22 17 21.1046 17 20V8C17 6.89543 16.1046 6 15 6H3C1.89543 6 1 6.89543 1 8V20C1 21.1046 1.89543 22 3 22Z"
                 stroke="#d5ffe9"
                 stroke-width="2"
+                class="svg"
               />
             </svg>
           </div>
