+++ frontend/gameUI/inputHandler.js
@@ -11,12 +11,7 @@ import gameState from "../gameFunctionality/gameState.js";
 import { hideExitPopup, showExitPopup } from "./uiController.js";
 import logger from "../gameFunctionality/logger.js";
 import { copyToClipboard } from "../gameUtils/copyToClipboard.js";
-import {
-  INPUT_CODES,
-  DEV_KEYWORDS,
-  MESSAGES,
-  INPUT_KEYS,
-} from "../gameConfig/inputConfig.js";
+import { INPUT, DEV_KEYWORDS, UI } from "../gameConfig.js";
 
 let keyword = "";
 let newSkin = null;
@@ -31,20 +26,20 @@ export function handleKeyDown(event) {
   const code = event.code;
   const key = event.key;
 
-  if (key === INPUT_KEYS.SPEEDUP) {
+  if (key === INPUT.KEYS.SPEEDUP) {
     handleSpeedup();
-  } else if (code === INPUT_CODES.FEED) {
+  } else if (code === INPUT.CODES.FEED) {
     handleFeed();
-  } else if (code === INPUT_CODES.MASK_SKIN && newSkin) {
+  } else if (code === INPUT.CODES.MASK_SKIN && newSkin) {
     handleMasking(newSkin);
-  } else if (code === INPUT_CODES.RESET_SKIN && newSkin) {
+  } else if (code === INPUT.CODES.RESET_SKIN && newSkin) {
     handleSkinReset();
-  } else if (code === INPUT_CODES.CLEAR_KEYWORD) {
+  } else if (code === INPUT.CODES.CLEAR_KEYWORD) {
     keyword = "";
-  } else if (code === INPUT_CODES.TOGGLE_PAUSE) {
+  } else if (code === INPUT.CODES.TOGGLE_PAUSE) {
     gameState.inactive = !gameState.inactive;
     gameState.inactive ? showExitPopup() : hideExitPopup();
-  } else if (code === INPUT_CODES.SPLIT) {
+  } else if (code === INPUT.CODES.SPLIT) {
     handleSplit();
   } else {
     keyword += key;
@@ -86,7 +81,7 @@ export function handleCanvasClick(event) {
       const distance = Math.sqrt(dx * dx + dy * dy);
       if (distance < cell.radius) {
         newSkin = gameState.playersSkins.find((p) => p.id === player.id)?.image;
-        copyToClipboard(() => player.name, MESSAGES.SKIN_COPY_TOOLTIP)();
+        copyToClipboard(() => player.name, UI.MESSAGES.SKIN_COPY_TOOLTIP)();
         return;
       }
     }
